Corregir tipo de dato BigDecimal en getOrderDetails

+++ src/main/java/co/edu/sena/productsreact/service/PaymentService.java
@@ -124,7 +124,7 @@ public class PaymentService {
                     r.getProduct().getId(),
                     r.getProduct().getNombre(),
                     r.getQuantity(),
-                    r.getProduct().getPrecio()
+                    r.getProduct().getPrecio().doubleValue()
                 ))
                 .toList();
